fix(stock): Move generic method to core utils
