+++ bloomstack_core/hook_events/item.py
@@ -1,8 +1,7 @@
 import json
 
-from six import string_types
-
 import frappe
+from bloomstack_core.utils import get_abbr
 from erpnext import get_default_company
 from erpnext.accounts.utils import get_company_default
 
@@ -44,43 +43,3 @@ def autoname(item, method=None):
 	item.name = item.item_code = item_code
 
 	return item_code
-
-
-def get_abbr(txt, max_length=2):
-	"""
-		Extract abbreviation from the given string as:
-			- Single-word strings abbreviate to the letters of the string, upto the max length
-			- Multi-word strings abbreviate to the initials of each word, upto the max length
-
-	Args:
-		txt (str): The string to abbreviate
-		max_length (int, optional): The max length of the abbreviation. Defaults to 2.
-
-	Returns:
-		str: The abbreviated string, in uppercase
-	"""
-
-	if not txt:
-		return
-
-	if not isinstance(txt, string_types):
-		try:
-			txt = str(txt)
-		except:
-			return
-
-	abbr = ""
-	words = txt.split(" ")
-
-	if len(words) > 1:
-		for word in words:
-			if len(abbr) >= max_length:
-				break
-
-			if word.strip():
-				abbr += word.strip()[0]
-	else:
-		abbr = txt[:max_length]
-
-	abbr = abbr.upper()
-	return abbr
