+++ bloomstack_core/utils.py
@@ -2,6 +2,8 @@ from __future__ import unicode_literals
 
 import json
 
+from six import string_types
+
 import frappe
 from erpnext.stock.doctype.batch.batch import get_batch_qty
 from python_metrc import METRC
@@ -29,6 +31,46 @@ def login_as(user):
 		return True
 
 	return False
+
+
+def get_abbr(txt, max_length=2):
+	"""
+		Extract abbreviation from the given string as:
+			- Single-word strings abbreviate to the letters of the string, upto the max length
+			- Multi-word strings abbreviate to the initials of each word, upto the max length
+
+	Args:
+		txt (str): The string to abbreviate
+		max_length (int, optional): The max length of the abbreviation. Defaults to 2.
+
+	Returns:
+		str: The abbreviated string, in uppercase
+	"""
+
+	if not txt:
+		return
+
+	if not isinstance(txt, string_types):
+		try:
+			txt = str(txt)
+		except:
+			return
+
+	abbr = ""
+	words = txt.split(" ")
+
+	if len(words) > 1:
+		for word in words:
+			if len(abbr) >= max_length:
+				break
+
+			if word.strip():
+				abbr += word.strip()[0]
+	else:
+		abbr = txt[:max_length]
+
+	abbr = abbr.upper()
+	return abbr
 
 
 @frappe.whitelist()
